requirements.txt
black==22.8.0
isort==5.10.1
pillow==5.1.0
